+++ mitra_models/src/filter_rules/types.rs
@@ -8,14 +8,14 @@ use crate::database::{
 #[derive(Clone, Copy, Debug, PartialEq)]
 pub enum FilterAction {
     Reject,
-    RejectMedia,
+    RejectMediaAttachments,
 }
 
 impl From<FilterAction> for i16 {
     fn from(value: FilterAction) -> i16 {
         match value {
             FilterAction::Reject => 1,
-            FilterAction::RejectMedia => 2,
+            FilterAction::RejectMediaAttachments => 2,
         }
     }
 }
@@ -26,7 +26,7 @@ impl TryFrom<i16> for FilterAction {
     fn try_from(value: i16) -> Result<Self, Self::Error> {
         let action = match value {
             1 => Self::Reject,
-            2 => Self::RejectMedia,
+            2 => Self::RejectMediaAttachments,
             _ => return Err(DatabaseTypeError),
         };
         Ok(action)